// src/app/modules/componentes-comunes/interfaces/reporte.interface.ts
export interface ReporteInterface {
    pesoTotal: number;
    cantidadCuentas: number;
    nombreAgricultor: string;
    emailAgricultor: string;
    usuarioAgricultor: string;
    telefonoAgricultor: string;
    toleranciaMenor: number; 
    toleranciaMayor: number;
    toleranciaCoincide: number;
}